chore(sendinblue): move sdk loader to a separate file
