+++ src/integrations/Sendinblue/browser.js
@@ -2,11 +2,11 @@
 /* eslint-disable class-methods-use-this */
 import logger from '../../utils/logUtil';
 import { NAME } from './constants';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { prepareUserTraits, prepareTrackEventData, preparePagePayload } from './utils';
 import { validateEmail, validatePhoneWithCountryCode } from '../../utils/commonUtils';
 
 import { getDefinedTraits } from '../../utils/utils';
+import { loadNativeSdk } from './nativeSdkLoader';
 
 class Sendinblue {
   constructor(config, analytics, destinationInfo) {
@@ -25,36 +25,7 @@ class Sendinblue {
 
   loadScript() {
     const { clientKey } = this;
-    (function () {
-      window.sib = {
-        equeue: [],
-        client_key: clientKey,
-      };
-      window.sendinblue = {};
-      for (var j = ['track', 'identify', 'trackLink', 'page'], i = 0; i < j.length; i++) {
-        (function (k) {
-          window.sendinblue[k] = function () {
-            var arg = Array.prototype.slice.call(arguments);
-            (
-              window.sib[k] ||
-              function () {
-                var t = {};
-                t[k] = arg;
-                window.sib.equeue.push(t);
-              }
-            )(arg[0], arg[1], arg[2], arg[3]);
-          };
-        })(j[i]);
-      }
-      var n = document.createElement('script'),
-        i = document.getElementsByTagName('script')[0];
-      (n.type = 'text/javascript'),
-        (n.id = 'sendinblue-js'),
-        (n.async = !0),
-        (n.src = 'https://sibautomation.com/sa.js?key=' + clientKey),
-        n.setAttribute('data-loader', LOAD_ORIGIN),
-        i.parentNode.insertBefore(n, i);
-    })();
+    loadNativeSdk(clientKey);
   }
 
   init() {
